code/requirements.txt
absl-py==0.12.0
blis==0.7.4
cachetools==4.2.1
catalogue==1.0.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
cymem==2.0.5
google-auth==1.28.0
google-auth-oauthlib==0.4.3
grpcio==1.36.1
idna==2.10
importlib-metadata==3.7.3
joblib==1.0.1
Markdown==3.3.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
murmurhash==1.0.5
nltk==3.6.2
oauthlib==3.1.0
olefile==0.46
plac==1.1.3
positional-encodings==3.0.0
preshed==3.0.5
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-video==1.1.11
scipy==1.5.4
spacy==2.3.5
srsly==1.0.5
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
thinc==7.4.5
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.26.4
wasabi==0.8.2
Werkzeug==1.0.1
zipp==3.4.1
